I've been faced with an endless loading page problem while working on the map
